# api/openapi.yaml
openapi: 3.1.0
info:
  title: Kyber API
  version: 0.4.0
  description: |
    Self-hosted issue & project tracker. Authenticate with the `kyber_session` cookie (browsers)
    or `Authorization: Bearer <token>` (API clients). Cookie-authenticated writes must send
    `Content-Type: application/json` (CSRF guard). Non-members receive `404` for project resources.
  license: { name: MIT, identifier: MIT }
servers:
  - url: http://localhost:8080
security:
  - bearer: []
  - cookie: []

paths:
  /healthz:
    get:
      operationId: healthz
      security: []
      responses:
        "200": { $ref: "#/components/responses/Status" }
  /readyz:
    get:
      operationId: readyz
      security: []
      responses:
        "200": { $ref: "#/components/responses/Status" }
        "503": { $ref: "#/components/responses/Status" }
  /metrics:
    get:
      operationId: metrics
      security: []
      responses:
        "200":
          description: Prometheus text exposition format.
          content:
            text/plain: { schema: { type: string } }

  /api/v1/auth/signup:
    post:
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email: { type: string }
                name: { type: string }
                password: { type: string, minLength: 10, maxLength: 128 }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Logged in; also sets the `kyber_session` cookie.
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties: { token: { type: string } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "429":
          description: Too many failed attempts for this email and IP.
          headers:
            Retry-After: { schema: { type: integer }, description: Seconds until the next attempt is allowed. }
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Error" } } }
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses:
        "204": { description: Session ended. }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/me:
    get:
      operationId: me
      responses:
        "200": { description: Current user, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "401": { $ref: "#/components/responses/Error" }

  /api/v1/projects:
    get:
      operationId: listProjects
      description: Projects the caller is a member of.
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ProjectList" } } } }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, name]
              properties:
                key: { $ref: "#/components/schemas/ProjectKey" }
                name: { type: string }
      responses:
        "201": { description: Created; the caller becomes admin., content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{key}:
    parameters: [{ $ref: "#/components/parameters/ProjectKey" }]
    get:
      operationId: getProject
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{key}/members:
    parameters: [{ $ref: "#/components/parameters/ProjectKey" }]
    get:
      operationId: listMembers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/Member" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: setMember
      description: Add a user by email or change their role (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "204": { description: Membership saved. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{key}/issues:
    parameters: [{ $ref: "#/components/parameters/ProjectKey" }]
    get:
      operationId: listIssues
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/Status" } }
      responses:
        "200": { description: Ordered by number., content: { application/json: { schema: { $ref: "#/components/schemas/IssueList" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type]
              properties:
                title: { type: string }
                type: { $ref: "#/components/schemas/IssueType" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/issues/{issueKey}:
    parameters: [{ $ref: "#/components/parameters/IssueKey" }]
    get:
      operationId: getIssue
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: editIssue
      description: Partial update guarded by optimistic locking. `assignee_id` null unassigns; omitted fields are unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: integer }
                title: { type: string }
                description: { type: string, maxLength: 20000 }
                priority: { $ref: "#/components/schemas/Priority" }
                assignee_id: { type: [string, "null"], format: uuid }
      responses:
        "200": { description: Updated issue, content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/issues/{issueKey}/transitions:
    parameters: [{ $ref: "#/components/parameters/IssueKey" }]
    post:
      operationId: transitionIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties: { to: { $ref: "#/components/schemas/Status" } }
      responses:
        "200": { description: Moved, content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/issues/{issueKey}/comments:
    parameters: [{ $ref: "#/components/parameters/IssueKey" }]
    get:
      operationId: listComments
      responses:
        "200":
          description: Oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/Comment" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: addComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties: { body: { type: string, maxLength: 10000 } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Comment" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }

components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: kyber_session }
  parameters:
    ProjectKey: { name: key, in: path, required: true, schema: { type: string } }
    IssueKey: { name: issueKey, in: path, required: true, schema: { type: string }, example: KYB-12 }
  responses:
    Error:
      description: RFC 9457 problem details; `title` is localized from Accept-Language (en, uz, ru).
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Error" } } }
    Status:
      description: Probe result
      content:
        application/json:
          schema:
            type: object
            required: [status]
            properties: { status: { type: string, enum: [ok, unavailable] } }
  schemas:
    Error:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string, description: Localized summary of the error code. }
        status: { type: integer }
        detail: { type: string, description: The specific reason for this occurrence. }
        instance: { type: string }
        code: { type: string, description: "Stable Kyber error code, e.g. ISSUE_NOT_FOUND. Branch on this." }
        numeric_code: { type: integer }
    User:
      type: object
      required: [id, email, name]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
    ProjectKey: { type: string, pattern: "^[A-Z][A-Z0-9]{1,9}$" }
    Project:
      type: object
      required: [id, key, name]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        key: { $ref: "#/components/schemas/ProjectKey" }
        name: { type: string }
    ProjectList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/Project" } } }
    Role: { type: string, enum: [admin, member, viewer] }
    Member:
      type: object
      required: [user_id, email, name, role]
      additionalProperties: false
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    Status: { type: string, enum: [todo, in_progress, done] }
    IssueType: { type: string, enum: [epic, story, task, bug, subtask] }
    Priority: { type: string, enum: [lowest, low, medium, high, highest] }
    Issue:
      type: object
      required: [id, key, title, type, status, description, priority, assignee_id, version]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        key: { type: string, pattern: "^[A-Z][A-Z0-9]{1,9}-[1-9][0-9]*$" }
        title: { type: string, minLength: 1 }
        type: { $ref: "#/components/schemas/IssueType" }
        status: { $ref: "#/components/schemas/Status" }
        description: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        assignee_id: { type: [string, "null"], format: uuid }
        version: { type: integer, minimum: 1 }
    IssueList:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/Issue" } } }
    Comment:
      type: object
      required: [id, author_id, author_name, body, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        author_id: { type: string, format: uuid }
        author_name: { type: string }
        body: { type: string, minLength: 1 }
        created_at: { type: string, format: date-time }